+++ src/Planes/View.py
@@ -60,22 +60,28 @@ class View(QTableView):
     
     @pyqtSlot()
     def update(self):
-        dia = Dialog(parent=self)
-        data = db.Plane(planeid=self.planeid).load()
-        dia.put(data)
-        if dia.exec():
-            dia.get(data)
-            data.save()
-            self.model().fresh()
+        if self.planeid:
+            dia = Dialog(parent=self)
+            data = db.Plane(planeid=self.planeid).load()
+            dia.put(data)
+            if dia.exec():
+                dia.get(data)
+                data.save()
+                self.model().fresh()
+        else:
+            QMessageBox.warning(self, 'Самолет', 'Сначала выберете нужный самолет')
     
     @pyqtSlot()
     def delete(self):
-        title = QApplication.translate('Planes.View', 'Plane')
-        q = QApplication.translate('Planes.View', 'Are you sure?')
-        ans = QMessageBox.question(self, title, q)
-        if ans == QMessageBox.Yes:
-            db.Plane(planeid=self.planeid).delete()
-            self.model().fresh()
+        if self.planeid:
+            title = QApplication.translate('Planes.View', 'Plane')
+            q = QApplication.translate('Planes.View', 'Are you sure?')
+            ans = QMessageBox.question(self, title, q)
+            if ans == QMessageBox.Yes:
+                db.Plane(planeid=self.planeid).delete()
+                self.model().fresh()
+        else:
+            QMessageBox.warning(self, 'Самолет', 'Сначала выберете нужный самолет')
     
     @pyqtSlot()
     def truncate(self):
